♻️ Refactor: ImgUploader Component
